Relocate some reusable multiint widget-related code
