+++ src/multiint.cpp
@@ -122,6 +122,7 @@
 #include "multivote.h"
 #include "campaigninfo.h"
 #include "screens/joiningscreen.h"
+#include "titleui/widgets/lobbyplayerrow.h"
 
 #include "activity.h"
 #include <algorithm>
@@ -205,19 +206,12 @@ static std::weak_ptr<WzMultiplayerOptionsTitleUI> currentMultiOptionsTitleUI;
 
 // widget functions
 static W_EDITBOX* addMultiEditBox(UDWORD formid, UDWORD id, UDWORD x, UDWORD y, char const *tip, char const *tipres, UDWORD icon, UDWORD iconhi, UDWORD iconid);
-static std::shared_ptr<W_FORM> createBlueForm(int x, int y, int w, int h, WIDGET_DISPLAY displayFunc = intDisplayFeBox);
 static W_FORM * addBlueForm(UDWORD parent, UDWORD id, UDWORD x, UDWORD y, UDWORD w, UDWORD h, WIDGET_DISPLAY displayFunc = intDisplayFeBox);
 static int numSlotsToBeDisplayed();
 static inline bool spectatorSlotsSupported();
-static inline bool isSpectatorOnlySlot(UDWORD playerIdx);
-//static inline bool isSpectatorOnlyPosition(UDWORD playerPosition);
 
 // Drawing Functions
 static void displayChatEdit(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
-static void displayPlayer(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
-static void displayReadyBoxContainer(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
-static void displayColour(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
-static void displayTeamChooser(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
 static void displaySpectatorAddButton(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
 static void displayAi(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
 static void displayDifficulty(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
@@ -225,19 +219,8 @@ static void displayMultiEditBox(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset
 static void drawBlueBox_Spectator(UDWORD x, UDWORD y, UDWORD w, UDWORD h);
 static void drawBlueBox_SpectatorOnly(UDWORD x, UDWORD y, UDWORD w, UDWORD h);
 void intDisplayFeBox_SpectatorOnly(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
-static inline void drawBoxForPlayerInfoSegment(UDWORD playerIdx, UDWORD x, UDWORD y, UDWORD w, UDWORD h);
 
 // pUserData structures used by drawing functions
-struct DisplayPlayerCache {
-	std::string	fullMainText;	// the “full” main text (used for storing the full player name when displaying a player)
-	WzText		wzMainText;		// the main text
-
-	std::string fullAltNameText;
-	WzText		wzAltNameText;
-
-	WzText		wzSubText;		// the sub text (used for players)
-	WzText		wzEloText;      // the elo text (used for players)
-};
 struct DisplayPositionCache {
 	WzText wzPositionText;
 };
@@ -261,42 +244,31 @@ static bool		SendPositionRequest(UBYTE player, UBYTE chosenPlayer);
 static bool		SendPlayerSlotTypeRequest(uint32_t player, bool isSpectator);
 bool changeReadyStatus(UBYTE player, bool bReady);
 static void stopJoining(std::shared_ptr<WzTitleUI> parent);
-static int difficultyIcon(int difficulty);
 
 static void sendRoomChatMessage(char const *text, bool skipLocalDisplay = false);
 
-static bool multiplayPlayersReady();
-static bool multiplayIsStartingGame();
 // ////////////////////////////////////////////////////////////////////////////
 // map previews..
 
-static const char *difficultyList[] = { N_("Super Easy"), N_("Easy"), N_("Medium"), N_("Hard"), N_("Insane") };
+static std::array<const char *, 5> difficultyList = { N_("Super Easy"), N_("Easy"), N_("Medium"), N_("Hard"), N_("Insane") };
 static const AIDifficulty difficultyValue[] = { AIDifficulty::SUPEREASY, AIDifficulty::EASY, AIDifficulty::MEDIUM, AIDifficulty::HARD, AIDifficulty::INSANE };
-static struct
-{
-	bool scavengers;
-	bool alliances;
-	bool teams;
-	bool power;
-	bool difficulty;
-	bool ai;
-	bool position;
-	bool bases;
-	bool spectators;
-} locked;
+static MultiplayOptionsLocked locked;
 static bool spectatorHost = false;
 static uint16_t defaultOpenSpectatorSlots = 0;
-
-struct AIDATA
-{
-	AIDATA() : name{0}, js{0}, tip{0}, difficultyTips{0}, assigned(0) {}
-	char name[MAX_LEN_AI_NAME];
-	char js[MAX_LEN_AI_NAME];
-	char tip[255 + 128];            ///< may contain optional AI tournament data
-	char difficultyTips[5][255];    ///< optional difficulty level info
-	int assigned;                   ///< How many AIs have we assigned of this type
-};
 static std::vector<AIDATA> aidata;
+
+const std::vector<AIDATA>& getAIData() { return aidata; }
+const MultiplayOptionsLocked& getLockedOptions() { return locked; }
+
+const char* getDifficultyListStr(size_t idx)
+{
+	ASSERT_OR_RETURN("", idx < difficultyList.size(), "Invalid idx: %zu", idx);
+	return difficultyList[idx];
+}
+size_t getDifficultyListCount()
+{
+	return difficultyList.size();
+}
 
 class ChatBoxButton : public W_BUTTON
 {
@@ -1115,16 +1087,6 @@ static std::shared_ptr<W_FORM> addInlineChooserBlueForm(const std::shared_ptr<W_
 	return std::dynamic_pointer_cast<W_FORM>(psForm->shared_from_this());
 }
 
-static std::shared_ptr<W_FORM> createBlueForm(int x, int y, int w, int h, WIDGET_DISPLAY displayFunc /* = intDisplayFeBox*/)
-{
-	ASSERT(displayFunc != nullptr, "Must have a display func!");
-	auto form = std::make_shared<W_FORM>();
-	form->setGeometry(x, y, w, h);
-	form->style = WFORM_PLAIN;
-	form->displayFunction = displayFunc;
-	return form;
-}
-
 static W_FORM * addBlueForm(UDWORD parent, UDWORD id, UDWORD x, UDWORD y, UDWORD w, UDWORD h, WIDGET_DISPLAY displayFunc /* = intDisplayFeBox*/)
 {
 	ASSERT(displayFunc != nullptr, "Must have a display func!");
@@ -1443,12 +1405,12 @@ static void addGameOptions()
 	updateLimitIcons();
 }
 
-static bool isHostOrAdmin()
+bool isHostOrAdmin()
 {
 	return NetPlay.isHost || NetPlay.players[selectedPlayer].isAdmin;
 }
 
-static bool isPlayerHostOrAdmin(uint32_t playerIdx)
+bool isPlayerHostOrAdmin(uint32_t playerIdx)
 {
 	ASSERT_OR_RETURN(false, playerIdx < MAX_CONNECTED_PLAYERS, "Invalid player idx: %" PRIu32, playerIdx);
 	return (playerIdx == NetPlay.hostPlayer) || NetPlay.players[playerIdx].isAdmin;
@@ -1494,7 +1456,7 @@ static int getPlayerTeam(int i)
  * Checks if all players are on the same team. If so, return that team; if not, return -1;
  * if there are no players, return team MAX_PLAYERS.
  */
-static int allPlayersOnSameTeam(int except)
+int allPlayersOnSameTeam(int except)
 {
 	int minTeam = MAX_PLAYERS, maxTeam = 0;
 	for (unsigned i = 0; i < game.maxPlayers; ++i)
@@ -2009,7 +1971,7 @@ public:
 	}
 };
 
-static std::set<uint32_t> validPlayerIdxTargetsForPlayerPositionMove(uint32_t player)
+std::set<uint32_t> validPlayerIdxTargetsForPlayerPositionMove(uint32_t player)
 {
 	std::set<uint32_t> validTargetPlayerIdx;
 	for (uint32_t i = 0; i < game.maxPlayers; i++)
@@ -2602,7 +2564,7 @@ bool recvTeamRequest(NETQUEUE queue)
 	return changeTeam(player, team, queue.index); // we do this regardless, in case of sync issues
 }
 
-static bool SendReadyRequest(UBYTE player, bool bReady)
+bool sendReadyRequest(UBYTE player, bool bReady)
 {
 	if (NetPlay.isHost)			// do or request the change.
 	{
@@ -3422,10 +3384,7 @@ static SwapPlayerIndexesResult recvSwapPlayerIndexes(NETQUEUE queue, const std::
 	return SwapPlayerIndexesResult::SUCCESS;
 }
 
-static bool canChooseTeamFor(int i)
-{
-	return (i == selectedPlayer || isHostOrAdmin());
-}
+
 
 // ////////////////////////////////////////////////////////////////////////////
 // tabs for player box
@@ -4098,488 +4057,6 @@ void WzPlayerBoxTabs::displayOptionsOverlay(const std::shared_ptr<WIDGET>& psPar
 
 	widgRegisterOverlayScreenOnTopOfScreen(optionsOverlayScreen, lockedScreen);
 }
-
-// ////////////////////////////////////////////////////////////////////////////
-// player row widgets
-
-class WzPlayerRow : public WIDGET
-{
-protected:
-	WzPlayerRow(uint32_t playerIdx, const std::shared_ptr<WzMultiplayerOptionsTitleUI>& parent)
-	: WIDGET()
-	, parentTitleUI(parent)
-	, playerIdx(playerIdx)
-	{ }
-
-public:
-	static std::shared_ptr<WzPlayerRow> make(uint32_t playerIdx, const std::shared_ptr<WzMultiplayerOptionsTitleUI>& parent)
-	{
-		class make_shared_enabler: public WzPlayerRow {
-		public:
-			make_shared_enabler(uint32_t playerIdx, const std::shared_ptr<WzMultiplayerOptionsTitleUI>& parent) : WzPlayerRow(playerIdx, parent) { }
-		};
-		auto widget = std::make_shared<make_shared_enabler>(playerIdx, parent);
-
-		std::weak_ptr<WzMultiplayerOptionsTitleUI> titleUI(parent);
-
-		// add team button (also displays the spectator "eye" for spectators)
-		widget->teamButton = std::make_shared<W_BUTTON>();
-		widget->teamButton->setGeometry(0, 0, MULTIOP_TEAMSWIDTH, MULTIOP_TEAMSHEIGHT);
-		widget->teamButton->UserData = playerIdx;
-		widget->teamButton->displayFunction = displayTeamChooser;
-		widget->attach(widget->teamButton);
-		widget->teamButton->addOnClickHandler([playerIdx, titleUI](W_BUTTON& button){
-			auto strongTitleUI = titleUI.lock();
-			ASSERT_OR_RETURN(, strongTitleUI != nullptr, "Title UI is gone?");
-			if ((!locked.teams || !locked.spectators))  // Clicked on a team chooser
-			{
-				if (canChooseTeamFor(playerIdx))
-				{
-					widgScheduleTask([strongTitleUI, playerIdx] {
-						strongTitleUI->openTeamChooser(playerIdx);
-					});
-				}
-			}
-		});
-
-		// add player colour
-		widget->colorButton = std::make_shared<W_BUTTON>();
-		widget->colorButton->setGeometry(MULTIOP_TEAMSWIDTH, 0, MULTIOP_COLOUR_WIDTH, MULTIOP_PLAYERHEIGHT);
-		widget->colorButton->UserData = playerIdx;
-		widget->colorButton->displayFunction = displayColour;
-		widget->attach(widget->colorButton);
-		widget->colorButton->addOnClickHandler([playerIdx, titleUI](W_BUTTON& button){
-			auto strongTitleUI = titleUI.lock();
-			ASSERT_OR_RETURN(, strongTitleUI != nullptr, "Title UI is gone?");
-			if (playerIdx == selectedPlayer || isHostOrAdmin())
-			{
-				if (!NetPlay.players[playerIdx].isSpectator) // not a spectator
-				{
-					widgScheduleTask([strongTitleUI, playerIdx] {
-						strongTitleUI->openColourChooser(playerIdx);
-					});
-				}
-			}
-		});
-
-		// add ready button
-		widget->updateReadyButton();
-
-		// add player info box (takes up the rest of the space in the middle)
-		widget->playerInfo = std::make_shared<W_BUTTON>();
-		widget->playerInfo->UserData = playerIdx;
-		widget->playerInfo->displayFunction = displayPlayer;
-		widget->playerInfo->pUserData = new DisplayPlayerCache();
-		widget->playerInfo->setOnDelete([](WIDGET *psWidget) {
-			assert(psWidget->pUserData != nullptr);
-			delete static_cast<DisplayPlayerCache *>(psWidget->pUserData);
-			psWidget->pUserData = nullptr;
-		});
-		widget->attach(widget->playerInfo);
-		widget->playerInfo->setCalcLayout([](WIDGET *psWidget) {
-			auto psParent = std::dynamic_pointer_cast<WzPlayerRow>(psWidget->parent());
-			ASSERT_OR_RETURN(, psParent != nullptr, "Null parent");
-			int x0 = MULTIOP_TEAMSWIDTH + MULTIOP_COLOUR_WIDTH;
-			int width = psParent->readyButtonContainer->x() - x0;
-			psWidget->setGeometry(x0, 0, width, psParent->height());
-		});
-		widget->playerInfo->addOnClickHandler([playerIdx, titleUI](W_BUTTON& button){
-			auto strongTitleUI = titleUI.lock();
-			ASSERT_OR_RETURN(, strongTitleUI != nullptr, "Title UI is gone?");
-			if (playerIdx == selectedPlayer || isHostOrAdmin())
-			{
-				uint32_t player = playerIdx;
-				// host/admin can move any player, clients can request to move themselves if there are available slots
-				if (((player == selectedPlayer && validPlayerIdxTargetsForPlayerPositionMove(player).size() > 0) ||
-						(NetPlay.players[player].allocated && isHostOrAdmin()))
-					&& !locked.position
-					&& player < MAX_PLAYERS
-					&& !isSpectatorOnlySlot(player))
-				{
-					widgScheduleTask([strongTitleUI, player] {
-						strongTitleUI->openPositionChooser(player);
-					});
-				}
-				else if (!NetPlay.players[player].allocated && !locked.ai && NetPlay.isHost)
-				{
-					if (button.getOnClickButtonPressed() == WKEY_SECONDARY && player < MAX_PLAYERS)
-					{
-						// Right clicking distributes selected AI's type and difficulty to all other AIs
-						for (int i = 0; i < MAX_PLAYERS; ++i)
-						{
-							// Don't change open/closed slots or humans or spectator-only slots
-							if (NetPlay.players[i].ai >= 0 && i != player && !isHumanPlayer(i) && !isSpectatorOnlySlot(i))
-							{
-								NetPlay.players[i].ai = NetPlay.players[player].ai;
-								NetPlay.players[i].isSpectator = NetPlay.players[player].isSpectator;
-								NetPlay.players[i].difficulty = NetPlay.players[player].difficulty;
-								setPlayerName(i, getAIName(player));
-								NETBroadcastPlayerInfo(i);
-							}
-						}
-						widgScheduleTask([strongTitleUI] {
-							strongTitleUI->updatePlayers();
-							resetReadyStatus(false);
-						});
-					}
-					else
-					{
-						widgScheduleTask([strongTitleUI, player] {
-							strongTitleUI->openAiChooser(player);
-						});
-					}
-				}
-			}
-		});
-
-		// update tooltips and such
-		widget->updateState();
-
-		return widget;
-	}
-
-	void geometryChanged() override
-	{
-		if (readyButtonContainer)
-		{
-			readyButtonContainer->callCalcLayout();
-		}
-		if (playerInfo)
-		{
-			playerInfo->callCalcLayout();
-		}
-	}
-
-	void updateState()
-	{
-		// update team button tooltip
-		if (NetPlay.players[playerIdx].isSpectator)
-		{
-			teamButton->setTip(_("Spectator"));
-		}
-		else if (canChooseTeamFor(playerIdx) && !locked.teams)
-		{
-			teamButton->setTip(_("Choose Team"));
-		}
-		else if (locked.teams)
-		{
-			teamButton->setTip(_("Teams locked"));
-		}
-		else
-		{
-			teamButton->setTip(nullptr);
-		}
-
-		// hide team button if needed
-		bool trueMultiplayerMode = (bMultiPlayer && NetPlay.bComms) || (!NetPlay.isHost && ingame.localJoiningInProgress);
-		if (!alliancesSetTeamsBeforeGame(game.alliance) && !NetPlay.players[playerIdx].isSpectator && !trueMultiplayerMode)
-		{
-			teamButton->hide();
-		}
-		else
-		{
-			teamButton->show();
-		}
-
-		// update color tooltip
-		if ((selectedPlayer == playerIdx || NetPlay.isHost) && (!NetPlay.players[playerIdx].isSpectator))
-		{
-			colorButton->setTip(_("Click to change player colour"));
-		}
-		else
-		{
-			colorButton->setTip(nullptr);
-		}
-
-		// update player info box tooltip
-		std::string playerInfoTooltip;
-		if ((selectedPlayer == playerIdx || NetPlay.isHost) && NetPlay.players[playerIdx].allocated && !locked.position && !isSpectatorOnlySlot(playerIdx))
-		{
-			playerInfoTooltip = _("Click to change player position");
-		}
-		else if (!NetPlay.players[playerIdx].allocated)
-		{
-			if (NetPlay.isHost && !locked.ai)
-			{
-				playerInfo->style |= WBUT_SECONDARY;
-				if (!isSpectatorOnlySlot(playerIdx))
-				{
-					playerInfoTooltip = _("Click to change AI, right click to distribute choice");
-				}
-				else
-				{
-					playerInfoTooltip = _("Click to close spectator slot");
-				}
-			}
-			else if (NetPlay.players[playerIdx].ai >= 0)
-			{
-				// show AI description. Useful for challenges.
-				playerInfoTooltip = aidata[NetPlay.players[playerIdx].ai].tip;
-			}
-		}
-		if (NetPlay.players[playerIdx].allocated)
-		{
-			const PLAYERSTATS& stats = getMultiStats(playerIdx);
-			if (!stats.identity.empty())
-			{
-				if (!playerInfoTooltip.empty())
-				{
-					playerInfoTooltip += "\n";
-				}
-				std::string hash = stats.identity.publicHashString(20);
-				playerInfoTooltip += _("Player ID: ");
-				playerInfoTooltip += hash.empty()? _("(none)") : hash;
-			}
-			std::string autoratingTooltipText;
-			if (stats.autorating.valid)
-			{
-				if (!stats.autorating.altName.empty())
-				{
-					if (!autoratingTooltipText.empty())
-					{
-						autoratingTooltipText += "\n";
-					}
-					std::string altnameStr = stats.autorating.altName;
-					if (altnameStr.size() > 128)
-					{
-						altnameStr = altnameStr.substr(0, 128);
-					}
-					size_t maxLinePos = nthOccurrenceOfChar(altnameStr, '\n', 1);
-					if (maxLinePos != std::string::npos)
-					{
-						altnameStr = altnameStr.substr(0, maxLinePos);
-					}
-					autoratingTooltipText += std::string(_("Alt Name:")) + " " + altnameStr;
-				}
-				if (!stats.autorating.details.empty()
-					&& stats.autoratingFrom == RATING_SOURCE_LOCAL) // do not display host-provided details (for now)
-				{
-					if (!autoratingTooltipText.empty())
-					{
-						autoratingTooltipText += "\n";
-					}
-					std::string detailsstr = stats.autorating.details;
-					if (detailsstr.size() > 512)
-					{
-						detailsstr = detailsstr.substr(0, 512);
-					}
-					size_t maxLinePos = nthOccurrenceOfChar(detailsstr, '\n', 10);
-					if (maxLinePos != std::string::npos)
-					{
-						detailsstr = detailsstr.substr(0, maxLinePos);
-					}
-					autoratingTooltipText += std::string(_("Player rating:")) + "\n" + detailsstr;
-				}
-			}
-			if (!autoratingTooltipText.empty())
-			{
-				if (!playerInfoTooltip.empty())
-				{
-					playerInfoTooltip += "\n\n";
-				}
-				if (stats.autoratingFrom == RATING_SOURCE_HOST)
-				{
-					playerInfoTooltip += std::string("[") + _("Host provided") + "]\n";
-				}
-				else
-				{
-					playerInfoTooltip += std::string("[") + astringf(_("From: %s"), getAutoratingUrl().c_str()) + "]\n";
-				}
-				playerInfoTooltip += autoratingTooltipText;
-			}
-		}
-		playerInfo->setTip(playerInfoTooltip);
-
-		// update ready button
-		updateReadyButton();
-	}
-
-public:
-
-	void updateReadyButton()
-	{
-		int disallow = allPlayersOnSameTeam(-1);
-
-		if (!readyButtonContainer)
-		{
-			// add form to hold 'ready' botton
-			readyButtonContainer = createBlueForm(MULTIOP_PLAYERWIDTH - MULTIOP_READY_WIDTH, 0,
-						MULTIOP_READY_WIDTH, MULTIOP_READY_HEIGHT, displayReadyBoxContainer);
-			readyButtonContainer->UserData = playerIdx;
-			attach(readyButtonContainer);
-			readyButtonContainer->setCalcLayout([](WIDGET *psWidget) {
-				auto psParent = psWidget->parent();
-				ASSERT_OR_RETURN(, psParent != nullptr, "Null parent");
-				psWidget->setGeometry(psParent->width() - MULTIOP_READY_WIDTH, 0, psWidget->width(), psWidget->height());
-			});
-		}
-
-		auto deleteExistingReadyButton = [this]() {
-			if (readyButton)
-			{
-				widgDelete(readyButton.get());
-				readyButton = nullptr;
-			}
-			if (readyTextLabel)
-			{
-				widgDelete(readyTextLabel.get());
-				readyTextLabel = nullptr;
-			}
-		};
-		auto deleteExistingDifficultyButton = [this]() {
-			if (difficultyChooserButton)
-			{
-				widgDelete(difficultyChooserButton.get());
-				difficultyChooserButton = nullptr;
-			}
-		};
-
-		if (!NetPlay.players[playerIdx].allocated && NetPlay.players[playerIdx].ai >= 0)
-		{
-			// Add AI difficulty chooser in place of normal "ready" button
-			deleteExistingReadyButton();
-			int playerDifficulty = static_cast<int8_t>(NetPlay.players[playerIdx].difficulty);
-			int icon = difficultyIcon(playerDifficulty);
-			char tooltip[128 + 255];
-			if (playerDifficulty >= static_cast<int>(AIDifficulty::EASY) && playerDifficulty < static_cast<int>(AIDifficulty::INSANE) + 1)
-			{
-				sstrcpy(tooltip, _(difficultyList[playerDifficulty]));
-				if (NetPlay.players[playerIdx].ai < aidata.size())
-				{
-					const char *difficultyTip = aidata[NetPlay.players[playerIdx].ai].difficultyTips[playerDifficulty];
-					if (strcmp(difficultyTip, "") != 0)
-					{
-						sstrcat(tooltip, "\n");
-						sstrcat(tooltip, difficultyTip);
-					}
-				}
-			}
-			bool freshDifficultyButton = (difficultyChooserButton == nullptr);
-			difficultyChooserButton = addMultiBut(*readyButtonContainer, MULTIOP_DIFFICULTY_INIT_START + playerIdx, 6, 4, MULTIOP_READY_WIDTH, MULTIOP_READY_HEIGHT,
-						(NetPlay.isHost && !locked.difficulty) ? _("Click to change difficulty") : tooltip, icon, icon, icon, MAX_PLAYERS, (NetPlay.isHost && !locked.difficulty) ? 255 : 125);
-			auto player = playerIdx;
-			auto weakTitleUi = parentTitleUI;
-			if (freshDifficultyButton)
-			{
-				difficultyChooserButton->addOnClickHandler([player, weakTitleUi](W_BUTTON&){
-					auto strongTitleUI = weakTitleUi.lock();
-					ASSERT_OR_RETURN(, strongTitleUI != nullptr, "Title UI is gone?");
-					if (!locked.difficulty && NetPlay.isHost)
-					{
-						widgScheduleTask([strongTitleUI, player] {
-							strongTitleUI->openDifficultyChooser(player);
-						});
-					}
-				});
-			}
-			return;
-		}
-		else if (!NetPlay.players[playerIdx].allocated)
-		{
-			// closed or open - remove ready / difficulty button
-			deleteExistingReadyButton();
-			deleteExistingDifficultyButton();
-			return;
-		}
-
-		if (disallow != -1)
-		{
-			// remove ready / difficulty button
-			deleteExistingReadyButton();
-			deleteExistingDifficultyButton();
-			return;
-		}
-
-		deleteExistingDifficultyButton();
-
-		bool isMe = playerIdx == selectedPlayer;
-		int isReady = NETgetDownloadProgress(playerIdx) != 100 ? 2 : NetPlay.players[playerIdx].ready ? 1 : 0;
-		char const *const toolTips[2][3] = {{_("Waiting for player"), _("Player is ready"), _("Player is downloading")}, {_("Click when ready"), _("Waiting for other players"), _("Waiting for download")}};
-		unsigned images[2][3] = {{IMAGE_CHECK_OFF, IMAGE_CHECK_ON, IMAGE_CHECK_DOWNLOAD}, {IMAGE_CHECK_OFF_HI, IMAGE_CHECK_ON_HI, IMAGE_CHECK_DOWNLOAD_HI}};
-
-		// draw 'ready' button
-		bool greyedOutReady = (NetPlay.players[playerIdx].isSpectator && NetPlay.players[playerIdx].ready) || (playerIdx != selectedPlayer);
-		bool freshReadyButton = (readyButton == nullptr);
-		readyButton = addMultiBut(*readyButtonContainer, MULTIOP_READY_START + playerIdx, 3, 10, MULTIOP_READY_WIDTH, MULTIOP_READY_HEIGHT,
-					toolTips[isMe][isReady], images[0][isReady], images[0][isReady], images[isMe][isReady], MAX_PLAYERS, (!greyedOutReady) ? 255 : 125);
-		ASSERT_OR_RETURN(, readyButton != nullptr, "Failed to create ready button");
-		readyButton->minClickInterval = GAME_TICKS_PER_SEC;
-		readyButton->unlock();
-		if (greyedOutReady && !NetPlay.isHost)
-		{
-			std::shared_ptr<WzMultiButton> pReadyBut_MultiButton = std::dynamic_pointer_cast<WzMultiButton>(readyButton);
-			if (pReadyBut_MultiButton)
-			{
-				pReadyBut_MultiButton->downStateMask = WBUT_DOWN | WBUT_CLICKLOCK;
-			}
-			auto currentState = readyButton->getState();
-			readyButton->setState(currentState | WBUT_LOCK);
-		}
-		if (freshReadyButton)
-		{
-			// must add onclick handler
-			auto player = playerIdx;
-			auto weakTitleUi = parentTitleUI;
-			readyButton->addOnClickHandler([player, weakTitleUi](W_BUTTON& button){
-				auto pButton = button.shared_from_this();
-				widgScheduleTask([weakTitleUi, player, pButton] {
-					auto strongTitleUI = weakTitleUi.lock();
-					ASSERT_OR_RETURN(, strongTitleUI != nullptr, "Title UI is gone?");
-
-					if (player == selectedPlayer
-						&& (!NetPlay.players[player].isSpectator || !NetPlay.players[player].ready || NetPlay.isHost)) // spectators can never toggle off "ready"
-					{
-						// Lock the "ready" button (until the request is processed)
-						pButton->setState(WBUT_LOCK);
-
-						SendReadyRequest(selectedPlayer, !NetPlay.players[player].ready);
-
-						// if hosting try to start the game if everyone is ready
-						if (NetPlay.isHost && multiplayPlayersReady())
-						{
-							startMultiplayerGame();
-							// reset flag in case people dropped/quit on join screen
-							NETsetPlayerConnectionStatus(CONNECTIONSTATUS_NORMAL, NET_ALL_PLAYERS);
-						}
-					}
-
-					if (NetPlay.isHost && !alliancesSetTeamsBeforeGame(game.alliance))
-					{
-						if (mouseDown(MOUSE_RMB) && player != NetPlay.hostPlayer) // both buttons....
-						{
-							std::string msg = astringf(_("The host has kicked %s from the game!"), getPlayerName(player));
-							sendRoomSystemMessage(msg.c_str());
-							kickPlayer(player, _("The host has kicked you from the game."), ERROR_KICKED, false);
-							resetReadyStatus(true);		//reset and send notification to all clients
-						}
-					}
-				});
-			});
-		}
-
-		if (!readyTextLabel)
-		{
-			readyTextLabel = std::make_shared<W_LABEL>();
-			readyButtonContainer->attach(readyTextLabel);
-			readyTextLabel->id = MULTIOP_READY_START + MAX_CONNECTED_PLAYERS + playerIdx;
-		}
-		readyTextLabel->setGeometry(0, 0, MULTIOP_READY_WIDTH, 17);
-		readyTextLabel->setTextAlignment(WLAB_ALIGNBOTTOM);
-		readyTextLabel->setFont(font_small, WZCOL_TEXT_BRIGHT);
-		readyTextLabel->setString(_("READY?"));
-	}
-
-private:
-	std::weak_ptr<WzMultiplayerOptionsTitleUI> parentTitleUI;
-	unsigned playerIdx = 0;
-	std::shared_ptr<W_BUTTON> teamButton;
-	std::shared_ptr<W_BUTTON> colorButton;
-	std::shared_ptr<W_BUTTON> playerInfo;
-	std::shared_ptr<WIDGET> readyButtonContainer;
-	std::shared_ptr<W_BUTTON> difficultyChooserButton;
-	std::shared_ptr<W_BUTTON> readyButton;
-	std::shared_ptr<W_LABEL> readyTextLabel;
-};
 
 // ////////////////////////////////////////////////////////////////////////////
 
@@ -5980,7 +5457,7 @@ static void loadMapPlayerSettings(WzConfig& ini)
 			}
 
 			WzString value = ini.value("difficulty", "Medium").toWzString();
-			for (unsigned j = 0; j < ARRAY_SIZE(difficultyList); ++j)
+			for (unsigned j = 0; j < difficultyList.size(); ++j)
 			{
 				if (strcasecmp(difficultyList[j], value.toUtf8().c_str()) == 0)
 				{
@@ -6754,7 +6231,7 @@ void handleAutoReadyRequest()
 	if (desiredReadyState != NetPlay.players[selectedPlayer].ready)
 	{
 		// Automatically set ready status
-		SendReadyRequest(selectedPlayer, desiredReadyState);
+		sendReadyRequest(selectedPlayer, desiredReadyState);
 	}
 }
 
@@ -7993,7 +7470,7 @@ void WzMultiplayerOptionsTitleUI::start()
 		{
 			processMultiopWidgets(MULTIOP_HOST);
 		}
-		SendReadyRequest(selectedPlayer, true);
+		sendReadyRequest(selectedPlayer, true);
 		if (getHostLaunch() == HostLaunch::Skirmish)
 		{
 			startMultiplayerGame();
@@ -8033,34 +7510,6 @@ void displayChatEdit(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
 
 // ////////////////////////////////////////////////////////////////////////////
 
-void displayTeamChooser(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
-{
-	int x = xOffset + psWidget->x();
-	int y = yOffset + psWidget->y();
-	UDWORD		i = psWidget->UserData;
-
-	ASSERT_OR_RETURN(, i < MAX_CONNECTED_PLAYERS, "Index (%" PRIu32 ") out of bounds", i);
-
-	drawBoxForPlayerInfoSegment(i, x, y, psWidget->width(), psWidget->height());
-
-	if (!NetPlay.players[i].isSpectator)
-	{
-		if (alliancesSetTeamsBeforeGame(game.alliance))
-		{
-			ASSERT_OR_RETURN(, NetPlay.players[i].team >= 0 && NetPlay.players[i].team < MAX_PLAYERS, "Team index out of bounds");
-			iV_DrawImage(FrontImages, IMAGE_TEAM0 + NetPlay.players[i].team, x + 1, y + 8);
-		}
-		else
-		{
-			// TODO: Maybe display something else here to signify "no team, FFA"
-		}
-	}
-	else
-	{
-		iV_DrawImage(FrontImages, IMAGE_SPECTATOR, x + 1, y + 8);
-	}
-}
-
 void displaySpectatorAddButton(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
 {
 	int x = xOffset + psWidget->x();
@@ -8070,7 +7519,7 @@ void displaySpectatorAddButton(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
 	iV_DrawImage(FrontImages, IMAGE_SPECTATOR, x + 2, y + 1);
 }
 
-static void displayAi(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
+void displayAi(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
 {
 	// Any widget using displayAi must have its pUserData initialized to a (DisplayAICache*)
 	assert(psWidget->pUserData != nullptr);
@@ -8083,6 +7532,7 @@ static void displayAi(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
 	drawBlueBox(x, y, psWidget->width(), psWidget->height());
 	WzString displayText;
 	PIELIGHT textColor = WZCOL_FORM_TEXT;
+
 	if (j >= 0)
 	{
 		if (j < aidata.size())
@@ -8117,7 +7567,7 @@ static void displayAi(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
 	cache.wzText.render(x + 10, y + 22, textColor);
 }
 
-static int difficultyIcon(int difficulty)
+int difficultyIcon(int difficulty)
 {
 	switch (difficulty)
 	{
@@ -8139,304 +7589,13 @@ static void displayDifficulty(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
 	const int y = yOffset + psWidget->y();
 	const int j = psWidget->UserData;
 
-	ASSERT_OR_RETURN(, j < ARRAY_SIZE(difficultyList), "Bad difficulty found: %d", j);
+	ASSERT_OR_RETURN(, j < difficultyList.size(), "Bad difficulty found: %d", j);
 	drawBlueBox(x, y, psWidget->width(), psWidget->height());
 	iV_DrawImage(FrontImages, difficultyIcon(j), x + 5, y + 5);
 	cache.wzDifficultyText.setText(gettext(difficultyList[j]), font_regular);
 	cache.wzDifficultyText.render(x + 42, y + 22, WZCOL_FORM_TEXT);
 }
 
-static bool isKnownPlayer(std::map<std::string, EcKey::Key> const &knownPlayers, std::string const &name, EcKey const &key)
-{
-	if (key.empty())
-	{
-		return false;
-	}
-	std::map<std::string, EcKey::Key>::const_iterator i = knownPlayers.find(name);
-	return i != knownPlayers.end() && key.toBytes(EcKey::Public) == i->second;
-}
-
-static void displayAltNameBox(int x, int y, WIDGET *psWidget, DisplayPlayerCache& cache, const PLAYERSTATS::Autorating& ar, bool isHighlight)
-{
-	int altNameBoxWidth = cache.wzAltNameText.width() + 4;
-	int altNameBoxHeight = cache.wzAltNameText.lineSize() + 2;
-	int altNameBoxX0 = (x + psWidget->width()) - altNameBoxWidth;
-	PIELIGHT altNameBoxColor = WZCOL_MENU_BORDER;
-	altNameBoxColor.byte.a = static_cast<uint8_t>(static_cast<float>(altNameBoxColor.byte.a) * (isHighlight ? 0.3f : 0.75f));
-	pie_UniTransBoxFill(altNameBoxX0, y, altNameBoxX0 + altNameBoxWidth, y + altNameBoxHeight, altNameBoxColor);
-
-	int altNameTextY0 = y + (altNameBoxHeight - cache.wzAltNameText.lineSize()) / 2 - cache.wzAltNameText.aboveBase();
-	PIELIGHT altNameTextColor = WZCOL_TEXT_MEDIUM;
-	if (ar.altNameTextColorOverride[0] != 255 || ar.altNameTextColorOverride[1] != 255 || ar.altNameTextColorOverride[2] != 255)
-	{
-		altNameTextColor = pal_Colour(ar.altNameTextColorOverride[0], ar.altNameTextColorOverride[1], ar.altNameTextColorOverride[2]);
-	}
-	if (isHighlight)
-	{
-		altNameTextColor.byte.a = static_cast<uint8_t>(static_cast<float>(altNameTextColor.byte.a) * 0.3f);
-	}
-	cache.wzAltNameText.render(altNameBoxX0 + 2, altNameTextY0, altNameTextColor);
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-void displayPlayer(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
-{
-	// Any widget using displayPlayer must have its pUserData initialized to a (DisplayPlayerCache*)
-	assert(psWidget->pUserData != nullptr);
-	DisplayPlayerCache& cache = *static_cast<DisplayPlayerCache *>(psWidget->pUserData);
-
-	int const x = xOffset + psWidget->x();
-	int const y = yOffset + psWidget->y();
-	unsigned const j = psWidget->UserData;
-	bool isHighlight = (psWidget->getState() & WBUT_HIGHLIGHT) != 0;
-
-	const int nameX = 32;
-
-	unsigned downloadProgress = NETgetDownloadProgress(j);
-
-	drawBoxForPlayerInfoSegment(j, x, y, psWidget->width(), psWidget->height());
-	if (downloadProgress != 100)
-	{
-		char progressString[MAX_STR_LENGTH];
-		ssprintf(progressString, j != selectedPlayer ? _("Sending Map: %u%% ") : _("Map: %u%% downloaded"), downloadProgress);
-		cache.wzMainText.setText(progressString, font_regular);
-		cache.wzMainText.render(x + 5, y + 22, WZCOL_FORM_TEXT);
-		cache.fullMainText = progressString;
-		return;
-	}
-	else if (ingame.localOptionsReceived && NetPlay.players[j].allocated)					// only draw if real player!
-	{
-		const PLAYERSTATS& stat = getMultiStats(j);
-		auto ar = stat.autorating;
-
-		std::string name = getPlayerName(j);
-
-		std::map<std::string, EcKey::Key> serverPlayers;  // TODO Fill this with players known to the server (needs implementing on the server, too). Currently useless.
-
-		PIELIGHT colour;
-		if (ingame.PingTimes[j] >= PING_LIMIT)
-		{
-			colour = WZCOL_FORM_PLAYER_NOPING;
-		}
-		else if (isKnownPlayer(serverPlayers, name, getMultiStats(j).identity))
-		{
-			colour = WZCOL_FORM_PLAYER_KNOWN_BY_SERVER;
-		}
-		else if (isLocallyKnownPlayer(name, getMultiStats(j).identity))
-		{
-			colour = WZCOL_FORM_PLAYER_KNOWN;
-		}
-		else
-		{
-			colour = WZCOL_FORM_PLAYER_UNKNOWN;
-		}
-
-		// name
-		if (cache.fullMainText != name)
-		{
-			if ((int)iV_GetTextWidth(name.c_str(), font_regular) > psWidget->width() - nameX)
-			{
-				while (!name.empty() && (int)iV_GetTextWidth((name + "...").c_str(), font_regular) > psWidget->width() - nameX)
-				{
-					name.resize(name.size() - 1);  // Clip name.
-				}
-				name += "...";
-			}
-			cache.wzMainText.setText(WzString::fromUtf8(name), font_regular);
-			cache.fullMainText = name;
-		}
-		WzString subText;
-		if (j == NetPlay.hostPlayer && NetPlay.bComms)
-		{
-			subText += _("HOST");
-		}
-		if (NetPlay.bComms && j != selectedPlayer)
-		{
-			char buf[250] = {'\0'};
-
-			// show "actual" ping time
-			ssprintf(buf, "%s%s: ", subText.isEmpty() ? "" : ", ", _("Ping"));
-			subText += buf;
-			if (ingame.PingTimes[j] < PING_LIMIT)
-			{
-				ssprintf(buf, "%03d", ingame.PingTimes[j]);
-			}
-			else
-			{
-				ssprintf(buf, "%s", "∞");  // Player has ping of somewhat questionable quality.
-			}
-			subText += buf;
-		}
-
-		if (!ar.valid)
-		{
-			ar.dummy = stat.played < 5;
-			// star 1 total droid kills
-			ar.star[0] = stat.totalKills > 600? 1 : stat.totalKills > 300? 2 : stat.totalKills > 150? 3 : 0;
-
-			// star 2 games played
-			ar.star[1] = stat.played > 200? 1 : stat.played > 100? 2 : stat.played > 50? 3 : 0;
-
-			// star 3 games won.
-			ar.star[2] = stat.wins > 80? 1 : stat.wins > 40? 2 : stat.wins > 10? 3 : 0;
-
-			// medals.
-			ar.medal = stat.wins >= 24 && stat.wins > 8 * stat.losses? 1 : stat.wins >= 12 && stat.wins > 4 * stat.losses? 2 : stat.wins >= 6 && stat.wins > 2 * stat.losses? 3 : 0;
-
-			ar.level = 0;
-			ar.autohoster = false;
-			ar.elo.clear();
-		}
-
-		if (cache.fullAltNameText != ar.altName)
-		{
-			std::string altName = ar.altName;
-			int maxAltNameWidth = static_cast<int>(static_cast<float>(psWidget->width() - nameX) * 0.65f);
-			iV_fonts fontID = font_small;
-			cache.wzAltNameText.setText(WzString::fromUtf8(altName), fontID);
-			cache.fullAltNameText = altName;
-			if (cache.wzAltNameText.width() > maxAltNameWidth)
-			{
-				while (!altName.empty() && ((int)iV_GetTextWidth(altName.c_str(), cache.wzAltNameText.getFontID()) + iV_GetEllipsisWidth(cache.wzAltNameText.getFontID())) > maxAltNameWidth)
-				{
-					altName.resize(altName.size() - 1);  // Clip alt name.
-				}
-				altName += "\u2026";
-				cache.wzAltNameText.setText(WzString::fromUtf8(altName), fontID);
-			}
-		}
-
-		if (!ar.altName.empty() && isHighlight)
-		{
-			// display first, behind everything
-			displayAltNameBox(x, y, psWidget, cache, ar, isHighlight);
-		}
-
-		int H = 5;
-		cache.wzMainText.render(x + nameX, y + 22 - H*!subText.isEmpty() - H*(ar.valid && !ar.elo.empty()), colour);
-		if (!subText.isEmpty())
-		{
-			cache.wzSubText.setText(subText, font_small);
-			cache.wzSubText.render(x + nameX, y + 28 - H*!ar.elo.empty(), WZCOL_TEXT_MEDIUM);
-		}
-
-		if (ar.autohoster)
-		{
-			iV_DrawImage(FrontImages, IMAGE_PLAYER_PC, x, y + 11);
-		}
-		else if (ar.dummy)
-		{
-			iV_DrawImage(FrontImages, IMAGE_MEDAL_DUMMY, x + 4, y + 13);
-		}
-		else
-		{
-			constexpr int starImgs[4] = {0, IMAGE_MULTIRANK1, IMAGE_MULTIRANK2, IMAGE_MULTIRANK3};
-			if (1 <= ar.star[0] && ar.star[0] < ARRAY_SIZE(starImgs))
-			{
-				iV_DrawImage(FrontImages, starImgs[ar.star[0]], x + 4, y + 3);
-			}
-			if (1 <= ar.star[1] && ar.star[1] < ARRAY_SIZE(starImgs))
-			{
-				iV_DrawImage(FrontImages, starImgs[ar.star[1]], x + 4, y + 13);
-			}
-			if (1 <= ar.star[2] && ar.star[2] < ARRAY_SIZE(starImgs))
-			{
-				iV_DrawImage(FrontImages, starImgs[ar.star[2]], x + 4, y + 23);
-			}
-			constexpr int medalImgs[4] = {0, IMAGE_MEDAL_GOLD, IMAGE_MEDAL_SILVER, IMAGE_MEDAL_BRONZE};
-			if (1 <= ar.medal && ar.medal < ARRAY_SIZE(medalImgs))
-			{
-				iV_DrawImage(FrontImages, medalImgs[ar.medal], x + 16 - 2*(ar.level != 0), y + 11);
-			}
-		}
-		constexpr int levelImgs[9] = {0, IMAGE_LEV_0, IMAGE_LEV_1, IMAGE_LEV_2, IMAGE_LEV_3, IMAGE_LEV_4, IMAGE_LEV_5, IMAGE_LEV_6, IMAGE_LEV_7};
-		if (ar.level > 0 && ar.level < ARRAY_SIZE(levelImgs))
-		{
-			iV_DrawImage(IntImages, levelImgs[ar.level], x + 24, y + 15);
-		}
-
-		if (!ar.elo.empty())
-		{
-			PIELIGHT eloColour = WZCOL_TEXT_BRIGHT;
-			if (ar.eloTextColorOverride[0] != 255 || ar.eloTextColorOverride[1] != 255 || ar.eloTextColorOverride[2] != 255)
-			{
-				eloColour = pal_Colour(ar.eloTextColorOverride[0], ar.eloTextColorOverride[1], ar.eloTextColorOverride[2]);
-			}
-			cache.wzEloText.setText(WzString::fromUtf8(ar.elo), font_small);
-			cache.wzEloText.render(x + nameX, y + 28 + H*!subText.isEmpty(), eloColour);
-		}
-
-		if (!ar.altName.empty() && !isHighlight)
-		{
-			// display last, over top of everything
-			displayAltNameBox(x, y, psWidget, cache, ar, isHighlight);
-		}
-	}
-	else	// AI
-	{
-		char aitext[100];
-
-		if (NetPlay.players[j].ai >= 0)
-		{
-			iV_DrawImage(FrontImages, IMAGE_PLAYER_PC, x, y + 11);
-		}
-
-		// challenges may use custom AIs that are not in aidata and set to 127
-		if (!challengeActive)
-		{
-		    ASSERT_OR_RETURN(, NetPlay.players[j].ai < (int)aidata.size(), "Uh-oh, AI index out of bounds");
-		}
-
-		PIELIGHT textColor = WZCOL_FORM_TEXT;
-		switch (NetPlay.players[j].ai)
-		{
-		case AI_OPEN:
-			if (!NetPlay.players[j].isSpectator)
-			{
-				sstrcpy(aitext, _("Open"));
-			}
-			else
-			{
-				sstrcpy(aitext, _("Spectator"));
-				textColor.byte.a = 220;
-			}
-			break;
-		case AI_CLOSED: sstrcpy(aitext, _("Closed")); break;
-		default: sstrcpy(aitext, getPlayerName(j)); break;
-		}
-		cache.wzMainText.setText(aitext, font_regular);
-		cache.wzMainText.render(x + nameX, y + 22, textColor);
-		cache.fullMainText = aitext;
-	}
-}
-
-static int factionIcon(FactionID faction)
-{
-	switch (faction)
-	{
-	case 0: return IMAGE_FACTION_NORMAL;
-	case 1: return IMAGE_FACTION_NEXUS;
-	case 2: return IMAGE_FACTION_COLLECTIVE;
-	default: return IMAGE_NO;	/// what??
-	}
-}
-
-void displayColour(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
-{
-	const int x = xOffset + psWidget->x();
-	const int y = yOffset + psWidget->y();
-	const int j = psWidget->UserData;
-
-	drawBoxForPlayerInfoSegment(j, x, y, psWidget->width(), psWidget->height());
-	if (!NetPlay.players[j].fileSendInProgress() && (isHumanPlayer(j) || NetPlay.players[j].difficulty != AIDifficulty::DISABLED) && !NetPlay.players[j].isSpectator)
-	{
-		int player = getPlayerColour(j);
-		STATIC_ASSERT(MAX_PLAYERS <= 16);
-		iV_DrawImageTc(FrontImages, IMAGE_PLAYERN, IMAGE_PLAYERN_TC, x + 3, y + 9, pal_GetTeamColour(player));
-		FactionID faction = NetPlay.players[j].faction;
-		iV_DrawImageFileAnisotropic(FrontImages, factionIcon(faction), x, y, Vector2f(11, 9));
-	}
-}
 
 // ////////////////////////////////////////////////////////////////////////////
 // Display blue box
@@ -8497,35 +7656,6 @@ void intDisplayFeBox_SpectatorOnly(WIDGET *psWidget, UDWORD xOffset, UDWORD yOff
 	int h = psWidget->height();
 
 	drawBlueBox_SpectatorOnly(x, y, w, h);
-}
-
-static inline void drawBoxForPlayerInfoSegment(UDWORD playerIdx, UDWORD x, UDWORD y, UDWORD w, UDWORD h)
-{
-	ASSERT(playerIdx < NetPlay.players.size(), "Invalid playerIdx: %zu", static_cast<size_t>(playerIdx));
-	if (playerIdx >= NetPlay.players.size() || !NetPlay.players[playerIdx].isSpectator)
-	{
-		drawBlueBox(x, y, w, h);
-	}
-	else
-	{
-		if (!isSpectatorOnlySlot(playerIdx))
-		{
-			drawBlueBox_Spectator(x, y, w, h);
-		}
-		else
-		{
-			drawBlueBox_SpectatorOnly(x, y, w, h);
-		}
-	}
-}
-
-static void displayReadyBoxContainer(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset)
-{
-	const int x = xOffset + psWidget->x();
-	const int y = yOffset + psWidget->y();
-	const int j = psWidget->UserData;
-
-	drawBoxForPlayerInfoSegment(j, x, y, psWidget->width(), psWidget->height());
 }
 
 // ////////////////////////////////////////////////////////////////////////////
@@ -8717,7 +7847,7 @@ std::shared_ptr<W_BUTTON> addMultiBut(const std::shared_ptr<W_SCREEN> &screen, U
 }
 
 /* Returns true if the multiplayer game can start (i.e. all players are ready) */
-static bool multiplayPlayersReady()
+bool multiplayPlayersReady()
 {
 	bool bReady = true;
 	size_t numReadyPlayers = 0;
@@ -8745,7 +7875,7 @@ static bool multiplayPlayersReady()
 	return bReady && (numReadyPlayers > 1) && (allPlayersOnSameTeam(-1) == -1);
 }
 
-static bool multiplayIsStartingGame()
+bool multiplayIsStartingGame()
 {
 	return bInActualHostedLobby && multiplayPlayersReady();
 }
@@ -8827,7 +7957,7 @@ static inline bool spectatorSlotsSupported()
 }
 
 // NOTE: Pass in the index in the NetPlay.players array
-static inline bool isSpectatorOnlySlot(UDWORD playerIdx)
+bool isSpectatorOnlySlot(UDWORD playerIdx)
 {
 	ASSERT_OR_RETURN(false, playerIdx < NetPlay.players.size(), "Invalid playerIdx: %" PRIu32 "", playerIdx);
 	return playerIdx >= MAX_PLAYERS || NetPlay.players[playerIdx].position >= game.maxPlayers;
